requirements.txt
discord==1.7.3
PyYAML==6.0
